Feature/restructure (#2)
* Rearranged file structure

* Current tests

* Added verified logo for pdf

* Changes to pdf generator

* Added content for pdf in assets/

* Modified spacing of input to 95%

* qr is now qr-image, pdf-kit is now html-pdf

* generator creates pdf upon camper POST
