+++ .gitignore
@@ -42,7 +42,10 @@ testem.log
 Thumbs.db
 
 # Generated QR pictures
-/QR
+/qr
 
 # Credentials
-/emailAdapter/credentials.json
+/config/credentials.json
+
+# Generated PDFs
+/pdf
