+++ api/routes/campers.js
@@ -9,7 +9,7 @@ router.use(function timeLog(req, res, next) {
 
 router.route('/')
   .get(campers.get_all_campers)
-  .post(campers.add_a_campers)
+  .post(campers.add_a_camper)
 router.route('/:camperId')
   .get(campers.view_a_camper)
   .put(campers.update_a_camper)
