+++ assets/pdf.html
@@ -1,0 +1,80 @@
+<style>
+  @font-face {
+    font-family: "Open Sans";
+    src: url("file:///app/assets/fonts/OpenSans-Regular.ttf")
+  }
+  @font-face {
+    font-family: "Open Sans";
+    src: url("file:///app/assets/fonts/OpenSans-Bold.ttf");
+    font-weight: bold
+  }
+  html,body {
+    font-family: "Open Sans";
+    font-size:10px;
+    text-align: justify;
+  }
+  p {
+    padding-bottom: 5px
+  }
+  .signature {
+    border: 0;
+    border-bottom: 1px solid #000;
+    margin: 0;
+  }
+</style>
+
+<div style="text-align: center">
+  <img src="file:///app/assets/img/camp-logo.svg" height=20px />
+</div>
+<div style="text-align: center">
+  <img src="file:///app/assets/img/pdf-verified-logo.svg" height=35px />
+</div>
+<p style="text-align:center; font-weight:bold">Permission / Agreement Form</p>
+<p>As parents / legal guardian of participant <b>FIRST_NAME LAST_NAME</b>,</p>
+<p>I understand and recognize that there may be inherent risks involved in the activities of <b>Camp LIVE: VERIFIED (High School Camp)</b> held under the auspices of Greenhills Christian Fellowship (GCF), from May 31-June 3, 2017, at Batangas Country Club, and I hereby release, discharge, and hold harmless GCF, its camp organizers, volunteers, employees, contractors or subcontractors (Organizers), including the officers or trustees of GCF, from any and all liability for injuries, including those that may result in death, and/or illnesses incurred while participating or attending in said Camp. I also give my expressed consent to the Organizers to transport said Participant to and from the Camp venue in church- owned or leased vehicles, and hold GCF and/or the Organizers free from any liability for any injury to or damage to the person of the Participant or to his/her belongings during travel.</p>
+<p>Moreover, in the event that I am not immediately available, should the Participant suffer a serious or life-threatening injury for which emergency medical treatment may be necessary, I hereby authorize the Organizers to engage qualified medical personnel to initiate any necessary medical treatment or care, and that they will use all reasonable efforts to notify me, where practical, and I hereby give permission to any such physician or other medical personnel to provide such medical treatment as deemed medically appropriate.</p>
+<p>By signing this document, the parent or legal guardian confirms that he or she has authority to sign, has read the entire document, and has understanding that the document waives certain rights as above indicated.</p>
+<p>
+  <div style="width:65%; display:table; float:left">
+    <label for="signature" style="display:table-cell; width:150px">Signature over Printed Name:</label>
+    <input type="text" id="signature" class="signature" style="display:table-cell; width:95%"/>
+  </div>
+  <div style="width:35%; display:table; float:right">
+    <label for="signature" style="display:table-cell; width:35px">Date:</label>
+    <input type="text" id="signature" class="signature" style="display:table-cell; width:95%"/>
+  </div>
+</p>
+<p>
+  <div style="width:60%; display:table; float:left">
+    <label for="signature" style="display:table-cell; width:135px">Relationship to Paricipant:</label>
+    <input type="text" id="signature" class="signature" style="display:table-cell; width:95%"/>
+  </div>
+  <div style="width:40%; display:table; float:right">
+    <label for="signature" style="display:table-cell; width:70px">Contact No.:</label>
+    <input type="text" id="signature" class="signature" style="display:table-cell; width:95%"/>
+  </div>
+</p>
+<div style="width:65%; float:left">
+  <p style="padding-bottom:7px">In compliance with the insurance company's requirements, we kindly ask you to fill out your parents' names and birthdates below.</p>
+  <p style="width:100%; display:table ">
+    <label for="signature" style="display:table-cell; width:85px">Father's Name:</label>
+    <input type="text" id="signature" class="signature" style="display:table-cell; width:95%"/>
+  </p>
+  <p style="width:100%; display:table">
+    <label for="signature" style="display:table-cell; width:125px">Birthday (MM/DD/YYYY):</label>
+    <input type="text" id="signature" class="signature" style="display:table-cell; width:95%"/>
+  </p>
+  <p style="width:100%; display:table; padding-top: 7px">
+    <label for="signature" style="display:table-cell; width:85px">Mother's Name:</label>
+    <input type="text" id="signature" class="signature" style="display:table-cell; width:95%"/>
+  </p>
+  <p style="width:100%; display:table">
+    <label for="signature" style="display:table-cell; width:125px">Birthday (MM/DD/YYYY):</label>
+    <input type="text" id="signature" class="signature" style="display:table-cell; width:95%"/>
+  </p>
+</div>
+<div style="width:35%; float:left; text-align: center">
+  <img src="file:///app/qr/CAMPER_ID.svg" width=60% style=" padding-top: 20%; padding-left: 20%" />
+  <div style="padding-left: 20%; font-size:8px">FIRST_NAME LAST_NAME</div>
+  <div style="padding-left: 20%; font-size:7.5px">CAMPER_ID</div>
+</div>
