--- qrAdapter/config.json
@@ -1,4 +1,0 @@
-{
-  "url": "api.qrserver.com",
-  "dir": "./QR"
-}
