+++ server.js
@@ -1,8 +1,8 @@
 var express = require('express'),
     app = express(),
-    port = process.env.PORT || 3000,
-    mongo_addr = process.env.MONGO_PORT_27017_TCP_ADDR,
-    mongo_port = process.env.MONGO_PORT_27017_TCP_PORT,
+    port = process.env.PORT || "3000",
+    mongo_addr = process.env.MONGO_PORT_27017_TCP_ADDR || "127.0.0.1",
+    mongo_port = process.env.MONGO_PORT_27017_TCP_PORT || "27017",
     mongoose = require('mongoose'),
     Person = require('./api/models/person'),
     Camper = require('./api/models/camper'),
